fix: fix test
