+++ tests/integrations/array/test_torch_train.py
@@ -3,6 +3,7 @@ from typing import Optional
 import torch
 
 from docarray import Document, DocumentArray
+from docarray.array.array_stacked import DocumentArrayStacked
 from docarray.typing import TorchTensor
 
 
@@ -14,9 +15,9 @@ def test_torch_train():
     N = 10
 
     batch = DocumentArray[Mmdoc](Mmdoc(text=f'hello{i}') for i in range(N))
-
     batch.tensor = torch.zeros(N, 3, 224, 224)
-    batch.stack()
+
+    batch = DocumentArrayStacked[Mmdoc](batch)
 
     class Model(torch.nn.Module):
         def __init__(self):
