+++ tests/units/array/test_stack_mod.py
@@ -1,14 +1,14 @@
-from typing import Optional
-
 import numpy as np
 import pytest
 import torch
 
 from docarray import Document, DocumentArray
+from docarray.array.array_stacked import DocumentArrayStacked
 from docarray.typing import NdArray, TorchTensor
 
 
-def test_stack():
+@pytest.fixture()
+def batch():
     class Image(Document):
         tensor: TorchTensor[3, 224, 224]
 
@@ -16,7 +16,55 @@ def test_stack():
         [Image(tensor=torch.zeros(3, 224, 224)) for _ in range(10)]
     )
 
-    batch.stack()
+    return DocumentArrayStacked[Image](batch)
+
+
+def test_len(batch):
+    assert len(batch) == 10
+
+
+def test_getitem(batch):
+    for i in range(len(batch)):
+        print(i)
+        assert (batch[i].tensor == torch.zeros(3, 224, 224)).all()
+
+
+def test_iterator(batch):
+    for doc in batch:
+        assert (doc.tensor == torch.zeros(3, 224, 224)).all()
+
+
+def test_stack_setter(batch):
+
+    batch.tensor = torch.ones(10, 3, 224, 224)
+
+    assert (batch.tensor == torch.ones(10, 3, 224, 224)).all()
+
+
+def test_stack_optional(batch):
+
+    assert (batch._columns['tensor'] == torch.zeros(10, 3, 224, 224)).all()
+    assert (batch.tensor == torch.zeros(10, 3, 224, 224)).all()
+
+
+def test_stack_numpy():
+    class Image(Document):
+        tensor: NdArray[3, 224, 224]
+
+    batch = DocumentArray[Image](
+        [Image(tensor=np.zeros((3, 224, 224))) for _ in range(10)]
+    )
+
+    batch = DocumentArrayStacked[Image](batch)
+
+    assert (batch._columns['tensor'] == np.zeros((10, 3, 224, 224))).all()
+    assert (batch.tensor == np.zeros((10, 3, 224, 224))).all()
+    assert batch.tensor.ctypes.data == batch._columns['tensor'].ctypes.data
+
+    batch.to_document_array(batch)
+
+
+def test_stack(batch):
 
     assert (batch._columns['tensor'] == torch.zeros(10, 3, 224, 224)).all()
     assert (batch.tensor == torch.zeros(10, 3, 224, 224)).all()
@@ -40,7 +88,7 @@ def test_stack_mod_nested_document():
         [MMdoc(img=Image(tensor=torch.zeros(3, 224, 224))) for _ in range(10)]
     )
 
-    batch.stack()
+    batch = DocumentArrayStacked[MMdoc](batch)
 
     assert (
         batch._columns['img']._columns['tensor'] == torch.zeros(10, 3, 224, 224)
@@ -54,7 +102,7 @@ def test_stack_mod_nested_document():
     )
 
 
-def test_unstack():
+def test_convert_to_da(batch):
     class Image(Document):
         tensor: TorchTensor[3, 224, 224]
 
@@ -62,11 +110,10 @@ def test_unstack():
         [Image(tensor=torch.zeros(3, 224, 224)) for _ in range(10)]
     )
 
-    batch.stack()
+    batch = DocumentArrayStacked[Image](batch)
+    da = batch.to_document_array(batch)
 
-    batch.unstack()
-
-    for doc in batch:
+    for doc in da:
         assert (doc.tensor == torch.zeros(3, 224, 224)).all()
 
 
@@ -81,98 +128,27 @@ def test_unstack_nested_document():
         [MMdoc(img=Image(tensor=torch.zeros(3, 224, 224))) for _ in range(10)]
     )
 
-    batch.stack()
+    batch = DocumentArrayStacked[MMdoc](batch)
 
-    batch.unstack()
+    da = batch.to_document_array(batch)
 
-
-def test_stack_runtime_error():
-    class Image(Document):
-        tensor: TorchTensor[3, 224, 224]
-
-    batch = DocumentArray[Image](
-        [Image(tensor=torch.zeros(3, 224, 224)) for _ in range(10)]
-    )
-
-    batch.stack()
-
-    with pytest.raises(RuntimeError):
-        batch.append([])
+    for doc in da:
+        assert (doc.img.tensor == torch.zeros(3, 224, 224)).all()
 
 
-def test_context_stack():
-    class Image(Document):
-        tensor: TorchTensor[3, 224, 224]
+def test_proto_stacked_mode_torch(batch):
 
-    batch = DocumentArray[Image](
-        [Image(tensor=torch.zeros(3, 224, 224)) for _ in range(10)]
-    )
-
-    assert not (batch.is_stacked())
-    with batch.stacked_mode():
-        assert batch.is_stacked()
-
-    assert not (batch.is_stacked())
+    batch.from_protobuf(batch.to_protobuf())
 
 
-def test_context_not_stack():
-    class Image(Document):
-        tensor: TorchTensor[3, 224, 224]
-
-    batch = DocumentArray[Image](
-        [Image(tensor=torch.zeros(3, 224, 224)) for _ in range(10)]
-    )
-
-    batch.stack()
-
-    assert batch.is_stacked()
-    with batch.unstacked_mode():
-        assert not (batch.is_stacked())
-
-    assert batch.is_stacked()
-
-
-def test_stack_setter():
-    class Image(Document):
-        tensor: TorchTensor[3, 224, 224]
-
-    batch = DocumentArray[Image](
-        [Image(tensor=torch.zeros(3, 224, 224)) for _ in range(10)]
-    )
-
-    batch.stack()
-
-    batch.tensor = torch.ones(10, 3, 224, 224)
-
-    assert (batch.tensor == torch.ones(10, 3, 224, 224)).all()
-
-
-def test_stack_optional():
-    class Image(Document):
-        tensor: Optional[TorchTensor[3, 224, 224]]
-
-    batch = DocumentArray[Image](
-        [Image(tensor=torch.zeros(3, 224, 224)) for _ in range(10)]
-    )
-
-    batch.stack()
-
-    assert (batch._columns['tensor'] == torch.zeros(10, 3, 224, 224)).all()
-    assert (batch.tensor == torch.zeros(10, 3, 224, 224)).all()
-
-
-def test_stack_numpy():
-    class Image(Document):
+def test_proto_stacked_mode_numpy():
+    class MyDoc(Document):
         tensor: NdArray[3, 224, 224]
 
-    batch = DocumentArray[Image](
-        [Image(tensor=np.zeros((3, 224, 224))) for _ in range(10)]
+    da = DocumentArray[MyDoc](
+        [MyDoc(tensor=np.zeros((3, 224, 224))) for _ in range(10)]
     )
 
-    batch.stack()
+    da = DocumentArrayStacked[MyDoc](da)
 
-    assert (batch._columns['tensor'] == np.zeros((10, 3, 224, 224))).all()
-    assert (batch.tensor == np.zeros((10, 3, 224, 224))).all()
-    assert batch.tensor.ctypes.data == batch._columns['tensor'].ctypes.data
-
-    batch.unstack()
+    da.from_protobuf(da.to_protobuf())
